+++ OMPParallelQuickSort.h
@@ -4,10 +4,10 @@ class OMPParallelQuickSort
 {
 public:
 	size_t size;
-	int* unsorted; // this has the sole purpose of displaying the result of the sort
-	int* sorted;
+	float* unsorted; // this has the sole purpose of displaying the result of the sort
+	float* sorted;
 
-	OMPParallelQuickSort(int* arr, size_t size);
+	OMPParallelQuickSort(float* arr, size_t size);
 	~OMPParallelQuickSort();
 
 private:
